% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.R
\name{coef.gnr}
\alias{coef.gnr}
\title{Print gross output function estimates}
\usage{
\method{coef}{gnr}(object, ...)
}
\arguments{
\item{object}{an object of class 'gnr'.}

\item{...}{currently not used.}
}
\value{
the named vector of parameter estimates contained in an object of class 'gnr'.
}
\description{
Print or return a numeric matrix of the estimated parameters from an object of class 'gnr'.
}
