% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colombian.R
\docType{data}
\name{colombian}
\alias{colombian}
\title{Colombian plant-level production data}
\format{
A data frame with seven variables:
\describe{
\item{\code{share}}{log intermediate input's revenue share}
\item{\code{id}}{firm id}
\item{\code{year}}{the last two digits of a year between 1981 and 1991}
\item{\code{RGO}}{log of real gross output with base year 1981}
\item{\code{L}}{log labor in level employee years}
\item{\code{K}}{log real capital stock with base year 1981}
\item{\code{RI}}{log real intermediate inputs with base year 1981}
}

For a complete listing of data, see \doi{10.1086/707736}.
}
\usage{
colombian
}
\description{
Plant-level production data for the food products industry (International Standard Industrial Classification code 311) in Colombia.
}
\references{
Gandhi, Amit, Salvador Navarro, and David Rivers. 2020. "On the Identification of Gross Output Production Functions." \emph{Journal of Political Economy}, 128(8): 2973-3016. \doi{10.1086/707736}.
}
\keyword{datasets}
