% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.R
\name{print.gnr}
\alias{print.gnr}
\title{Printing gross output function estimates}
\usage{
\method{print}{gnr}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object of class 'gnr'.}

\item{digits}{the number of significant figures to use for printing.}

\item{...}{currently not used.}
}
\value{
\code{print.gnr} has no return value and only prints a brief overview of elements contained in an object of class 'gnr' as described in the description.
}
\description{
Print estimates of the parameters in a gross output function and names of the output, input, and data from \code{\link[gnrprod]{gnrprod}}.
}
