% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.R
\name{print.gnrflex}
\alias{print.gnrflex}
\title{Printing first stage estimate}
\usage{
\method{print}{gnrflex}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object of class 'gnrflex'.}

\item{digits}{the number of significant figures to use for printing.}

\item{...}{currently not used.}
}
\value{
\code{print.gnrflex} has no return value and only prints a brief overview of elements contained in an object of class 'gnrflex' as described in the description.
}
\description{
Print estimate of the flexible input elasticity, the sum of squared residuals, and convergence status from \code{\link[gnrprod]{gnrflex}}.
}
