% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.R
\name{print.summary.gnr}
\alias{print.summary.gnr}
\title{Printing a summary of gross output function estimation}
\usage{
\method{print}{summary.gnr}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object of class 'summary.gnr'.}

\item{digits}{the number of significant figures to use for printing.}

\item{...}{currently not used.}
}
\value{
\code{print.gnr} has no return value and only prints the elements contained in an object of class 'summary.gnr' as described in the description.
}
\description{
Print a summary of the gross output function estimation routine from \code{\link[gnrprod]{gnrprod}}: names of output and inputs, summary statistics of the estimated productivity, the function estimates and standard errors if applicable, and convergence information.
}
